feat: Implement project management modals and functionality - 1 feat: Implement project management modals and functionality
- Added AddProjectBtn component for creating and editing projects.
- Introduced DeleteProjectModal for confirming project deletions.
- Created ProjectFormFields for handling project input fields.
- Developed ProjectImageUpload for managing project images.
- Implemented ProjectOptionsMenu for project actions (edit/delete).
- Added categories constant for project categorization.
