+++ packages/Frontend/src/App.jsx
@@ -9,8 +9,8 @@ import ResetPassword from "./component/pages/auth/ResetPassword";
 import Dashboard from "./component/pages/dashboard/Dashboard";
 import { ToastContainer } from "react-toastify";
 import Layout from "./shared/Layout";
-import Projects from "./component/pages/Projects/Projects";
-import ProjectDetails from "./component/pages/Projects/ProjectDetails";
+import Projects from "./component/pages/Projects/ProjectCard/Projects";
+import ProjectDetails from "./component/pages/Projects/ProjectDetails/ProjectDetails";
 import ProjectTasks from "./component/pages/Projects/ProjectTasks";
 import CompleteSigninGoogle from "./component/pages/auth/completeSigninGoogle";
 import GoogleCallback from "./component/pages/auth/GoogleCallback";
