+++ packages/Frontend/src/api/invite.js
@@ -35,4 +35,3 @@ export const sendInvite = async (userId, projectId, inviteData) => {
     inviteData
   );
 };
-// فايل هوك يشغل ملف ده زي ده
