+++ packages/Frontend/src/constants/categories.js
@@ -1,0 +1,71 @@
+export const categories = [
+  "UI/UX",
+  "Development",
+  "Marketing",
+  "Product Management",
+  "Business Strategy",
+  "Operations",
+  "Sales",
+  "Customer Support",
+  "Finance",
+  "Legal",
+  "Human Resources",
+  "IT Support",
+  "Research & Development",
+  "Data Science",
+  "Machine Learning",
+  "Artificial Intelligence",
+  "Cybersecurity",
+  "Blockchain",
+  "E-commerce",
+  "Supply Chain",
+  "Logistics",
+  "Event Management",
+  "Healthcare",
+  "Real Estate",
+  "Education",
+  "Non-Profit",
+  "Construction",
+  "Retail",
+  "Entertainment",
+  "Media & Publishing",
+  "Public Relations",
+  "Government",
+  "Telecommunications",
+  "Design",
+  "Consulting",
+  "Travel & Hospitality",
+  "Frontend Development",
+  "Backend Development",
+  "Fullstack Development",
+  "Quality Assurance",
+  "Testing",
+  "Bug Fixing",
+  "DevOps",
+  "Continuous Integration",
+  "Deployment",
+  "Cloud Computing",
+  "AWS",
+  "Google Cloud",
+  "Azure",
+  "Augmented Reality",
+  "Virtual Reality",
+  "AR/VR Development",
+  "Internet of Things",
+  "IoT Development",
+  "Smart Devices",
+  "Automation",
+  "Robotic Process Automation (RPA)",
+  "CRM Development",
+  "Customer Insights",
+  "CRM Integration",
+  "ERP Development",
+  "Enterprise Solutions",
+  "Business Systems",
+  "Game Development",
+  "Unity Development",
+  "Unreal Engine",
+  "AI Ethics",
+  "Ethical AI",
+  "AI Regulations",
+];
